Add shebang, fix compatibility with Prisma 5.7.0
The main script was missing a shebang, which caused it not to work with
several package manager and/or OS combinations. This adds that in, which
should fix the issues.

Additionally, the 5.7.0 version of @prisma/generator-helper no longer
seems to support named imports when importing from an ESM package, so
this switches to using the default export.

This also updates all dependencies to their latest versions.