Insert shebang in postbuild script
The shebang on the top of the TS file causes GitHub to identify it as
javascript, which kind of annoyed me. So this removes it from the TS
file, and adds it to the built JS file using a node command. Using node
here instead of something like perl since node should always be
installed if you're able to run build.