Fix 'object' enumType types
The enum types exported when using the 'object' enumType would always
just be a string, since the strings used in the enum object would be
widened to just be a 'string' type. This adds 'as const' to the
declaration, to prevent this widening.

See https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-4.html#const-assertions